requirements.txt
streamlit-1.40.1-py2.py3-none-any.whl